camera.service: Add CameraService with image and state handling
For multi-cam plugin, we have to isolate the camera on a dedicate ionic page so that each page has only one specific task to do for easier debugging. This CameraService is introduced to manage captured images and camera state using RxJS Observables. Included methods for adding, removing, and clearing images, as well as managing the capturing state.