camera: Implement CameraPage with routing and UI
This ports the implementation from the plugin demo and creates a CameraPage with its corresponding module and routing.

The tab #3 can then open this page using navCtrl.

The reason for such architecture of creating a new page instead of using tab#3 is that it makes sure that ionic doesn't override any styling related to transparency.

The page includes a UI for capturing images using the camera. Integrates CameraService for managing captured images and state.