camera: Update camera-multi-capture to version 0.0.5
Bumped the camera-multi-capture dependency to version 0.0.5. Refactored the CameraPage and CameraService to directly handle CameraImageData objects instead of URIs, improving the image management process. This change enhances the integration with the updated camera plugin. Now with the update, the CameraImageData also contains webPath and can be used by photo.service to save.